Backend!
I created a PHP script to handle our data.  It's pretty basic right now
and it doesn't totally work.  I'm familiar with mysql but not mysqli so
probably a few small errors.  Working to get it there though.  It's a
start.  Also renamed some things to make it flow better and work with
the item listing page