DB Works
You can now connect to the database when you run the code.  Nothing goes
in yet though.  About to add that.  I'll also add code to print the
items table so you can confirm your order went in.  Later I'll modify
this so that it brings you to your new listing page